rework to add explicit fourier sampler and generic mixture

+++ src/tree_world/models/fourier_metric.py
@@ -42,7 +42,7 @@ class FourierMetric(torch.nn.Module):
         c2s1 = location2[..., 0] * location1[..., 1].float()
         c2c1 = location2[..., 0] * location1[..., 0].float()
         s2s1 = location2[..., 1] * location1[..., 1].float()
-        delta_thetas = torch.atan2(s2c1 - c2s1, c2c1 + s2s1 + 1e-6).to(location1.dtype)
+        delta_thetas = torch.atan2(s2c1 - c2s1, c2c1 + s2s1 + 1e-6)
 
         # estimated displacements from thetas, made as small as possible solving across alphas -- but may not agree!
         # deltas has shape (..., J, d)
@@ -54,7 +54,7 @@ class FourierMetric(torch.nn.Module):
         )
         mean_deltas = deltas.mean(dim=-2)
 
-        return deltas, mean_deltas
+        return deltas.to(location1.dtype), mean_deltas.to(location1.dtype)
     
     def pseudo_distance(self, location1: torch.Tensor, location2: torch.Tensor, squared: bool=False):
         deltas, mean_deltas = self.compute_displacements(location1, location2)
